+++ packages/cli/src/agoras/cli/platforms/instagram.py
@@ -26,8 +26,8 @@ from argparse import ArgumentParser, Namespace, _SubParsersAction
 from agoras.platforms.instagram.wrapper import main as instagram_main
 
 from ..base import add_common_content_options
-from ..media_help import video_url_help
 from ..converter import ParameterConverter
+from ..media_help import video_url_help
 from ..validator import ActionValidator
 
 
@@ -151,7 +151,7 @@ def _add_video_options(parser: ArgumentParser):
     video.add_argument(
         '--video-type',
         metavar='<type>',
-        help='Video type (e.g., REELS, STORIES)'
+        help='Video type: REELS (default) or STORIES (case-insensitive; reel/story also accepted)'
     )